+++ packages/nimble-components/src/table-column/table-column-text.ts
@@ -1,6 +1,6 @@
 import { attr, html, ViewTemplate } from '@microsoft/fast-element';
 import { DesignSystem } from '@microsoft/fast-foundation';
-import type { StringField, TableCellState } from '../../types';
+import type { StringField, TableCellState } from '../table/types';
 import { TableColumn } from './table-column';
 
 type TableColumnTextCellData = StringField<'value'>;
@@ -13,9 +13,6 @@ interface TableColumnTextColumnConfig {
  */
 export class TableColumnText extends TableColumn<TableColumnTextCellData, TableColumnTextColumnConfig> {
     public cellStateDataFieldNames = ['value'] as const;
-
-    @attr({ attribute: 'column-id' })
-    public columnId?: string;
 
     @attr({ attribute: 'value-key' })
     public valueKey!: string;
